LocalValidationError and separation of errors
    - Move each error into its own file
    - Rename `ValidationError` to `LocalValidationError`
