+++ src/BadKeyError.ts
@@ -1,0 +1,6 @@
+export class BadKeyError extends Error {
+    public constructor() {
+        super("Failed to parse correct key");
+        this.name = "BadKeyError";
+    }
+}
