+++ src/crypto/Ed25519PrivateKey.ts
@@ -5,7 +5,7 @@ import { Mnemonic } from "./Mnemonic";
 import { decodeHex, deriveChildKey, ed25519PrivKeyPrefix, encodeHex, pbkdf2, randomBytes } from "./util";
 import { RawKeyPair } from "./RawKeyPair";
 import { createKeystore, loadKeystore } from "./Keystore";
-import { BadKeyError } from "../errors";
+import { BadKeyError } from "../BadKeyError";
 
 export class Ed25519PrivateKey {
     public readonly publicKey: Ed25519PublicKey;
